Added project files for Visual Studio.
triumph and tas now build on VS, and on Mac/Linux via CMake.

+++ triumph/include/Screen.h
@@ -18,8 +18,8 @@ private:
 	std::vector<uint32_t> byte_framebuffer;
 
 	// screen size
-	int const PIXELS_PER_TRIT = 1;
-	int const PIXELS_PER_TILE = 9 * PIXELS_PER_TRIT;
+	size_t const PIXELS_PER_TRIT = 1;
+	size_t const PIXELS_PER_TILE = 9 * PIXELS_PER_TRIT;
 	int const TILE_GRID_WIDTH = 54;
 	int const TILE_GRID_HEIGHT = 54;
 	int const PIXEL_WIDTH = TILE_GRID_WIDTH * PIXELS_PER_TILE;
